// admin-panel/src/api/admin/types.ts
export type CleanupQueryUserFile = {
  id: number;
  slug: string;
  userId: number;
  username: string;
  fileName: string;
  fileSize: number;
  createdAt: number;
};